requirements.txt
asgiref==3.7.2
crispy-bootstrap5==0.7
da-vinci==0.3.0
django-ajax-datatable==4.4.5
django-appconf==1.0.5
django-cleanup==8.0.0
django-crispy-forms==2.0
django-htmx==1.15.0
django-quill-editor==0.1.40
django-select2==8.1.2
django-thumbnails==0.7.0
django-user-agents==0.4.0
pendulum==2.1.2
Pillow==10.0.0
python-dateutil==2.8.2
pytz==2023.3
pytzdata==2020.1
shortuuid==1.0.11
six==1.16.0
sqlparse==0.4.4
ua-parser==0.16.1
user-agents==2.2.0
